+++ __tests__/pages/(dashboard)/layout.test.tsx
@@ -1,8 +1,8 @@
 import '@testing-library/jest-dom'
 import {render} from '@testing-library/react'
-import Layout from "@/app/\(dashboard\)/layout"
+import Layout from "@/app/(dashboard)/layout"
 describe('Layout tests',()=>{
     test('renders without crashing',()=>{
-        render(<Layout/>)
-    })
-})
+        render(<Layout/>);
+    });
+});
