+++ src/app/(dashboard)/layout.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react";
 
 const Layout:FC= ()=>{
-    return <div>foo</div>
+    return <div>foo</div>;
 }
 
-export default Layout
+export default Layout;
